upload sidemenu
